Allow flagging pipeline middleware as error handlers
Prior to this patch, if you specified pipeline middleware as a service
name, the closure returned prevented it from being used as error
middleware, as the arity was incorrect.

Since we cannot determine if a service is error middleware until after
retrieval, this patch adds a new flag to the pipeline middleware
specifications, `error`. When `true`, this will cause the closure used
for lazy-loading to have the error middleware signature (arity 4).