[#56] Allow non-array $params to render()
Some templating implementations, such as Mustache and zend-view, allow
passing an object to the renderer; the renderer then determines how this
is interpreted. In the case of zend-view, it pulls variables from the
view model in a specific way, and also allows access to the view model
from within the template. In the case of Mustache, certain features such
as higher-order functions (similar to helpers) are accomplished via
callable properties of the view object.

This patch modifies `TemplateInterface::render()` to allow arbitrary
types for the `$params` argument, while:

- stipulating in the docblock that only arrays or objects are acceptable
- providing an empty array as a default value

For the two current implementations, Plates and Twig, this introduces an
`ArrayParametersTrait` that will attempt to munge null and object values
to arrays, and raises an exception on other params types.