[#122] Import factories from expressive-composer-installer
This patch imports the factories introduced at:

- https://github.com/xtreamwayz/expressive-composer-installer/tree/master/src/App/Template

for each of Twig, Plates, and ZendView, with modifications (particularly the
lack of default template paths). Additionally, unit tests have been added to
verify behavior of each container.

The Twig support now includes the TwigExtension, which overrides the path and
asset twig helpers in order to delegate to zend-expressive functionality.