In Aura router, inject server params with request method
Per #140, to simplify testing, we should be able to use the PSR-7 server
request instance's request method, without having to also set it in the
server parameters. This patch updates the tests to indicate the
request's `getMethod()` method will be called, while its
`getServerParams()` returns an empty array. The `match()` method now
injects the returned parameters with the `REQUEST_METHOD` before passing
them on to the underlying router's matching functionality.