Moves the RouteResultObserverInterface back to the Expressive namespace
This patch moves the `RouteResultObserverInterface` into the `Zend\Expressive`
namespace. Since it is triggered by `Application`, it is an application-level
responsibility, not the router's.

Since the class names do not conflict, this can lead to an immediate deprecation
of the original interface in the zend-expressive-router package. However,
zend-expressive-zendviewrenderer will need updates so that it fulfills the new
interface via its `UrlHelper` instance.