Edited recipe
This patch edits the recipe to demonstrate the problem earlier, as well as
provide both a simple switch-based solution as well as the abstract solution.
The abstract solution was simplified to remove reflection usage; we can assume
the parameters for `Action` methods will accept what we provide them, and that
narrative was distracting from the goal of the recipe.

Finally, this also provides a trait-based solution as an alternative.