Simplify the middleware pipeline configuration
This patch simplifies the middleware pipeline configuration, by removing the
`pre_routing` and `post_routing` keys; this also allows developers to replace
the routing middleware if desired by omitting the default from the pipeline
configuration.

All items in the pipeline follow the same structure that was required of the
`pre_` and `post_routing` keys, with one exception: the routing middleware is
specified via a constant value.