Separate routing/dispatch middleware
This patch addresses #259 and comments from @danizord, separating
`routeMiddleware()` into two distinct methods, `routeMiddleware()` and
`dispatchMiddleware()`.

This change also means we can no longer auto-register the routing middleware
with the application. However, it enables new use cases, such as middleware
interceptors between routing and dispatch, as well as substituting alternate
routing and/or dispatch middleware in the application.

Because the application no longer auto-registers the routing/dispatch
middleware, the `ApplicationFactory` was updated to no longer raise an exception
in situations where routes are registered but the routing middleware is not
piped — since alternate routing middleware may be taking that responsibility.