Implemented priority for middleware specifications
This patch implements priority for middleware specifications provided to
configuration used by the `ApplicationFactory`. Each specification can now
include a `priority` key with an integer value; if none is provided, `1` is
assumed as the default. The `ApplicationFactory` uses the priority to seed an
`SplPriorityQueue` instance, which is then iterated when piping to the
application. The priority queue is seeded using similar semantics to those in
zend-stdlib, ensuring that middleware with the same priority are dequeued in the
same order of attachment.

The changes allow usage of the constant values indicating routing/dispatch/route
result observer middleware both as middleware pipeline values, as well as inside
`middleware` array specifications. These special values *always* default to
priority 1.

The migration path for developers is to add `priority` values to their
middleware specs; until that is done, we cannot guarantee order of merged
values!