Implement request/response decoration
Currently, the `Zend\Stratigility\Next` implementation requires that the
request and response passed to it are the Stratigility-specific
decorators. As such, the `ErrorMiddlewarePipe` is required to decorate
the incoming request and response. I chose to use reflection here so as
not to duplicate logic; ideally, we can remove that requirement from the
Stratigility library in the future.