Ensure error middleware pipelines can be executed
This patch provides tests for the following scenarios:

- Given a middleware spec that resolves to a middleware service and which
  specifies error middleware, the error middleware will be executed in error
  conditions.
- Given a middleware spec that resolves to an error middleware pipeline and
  which specifies error middleware, the error middleware pipeline will be
  executed in error conditions.

The latter was failing, prompting this patch.

The fix was reasonably simple: when an object is used as error middleware,
`Zend\Stratigility\Utils::getArity()` looks at the number of *required*
arguments, and omits *optional* arguments, despite the fact that the
`ErrorMiddlewareInterface` specifies the fourth argument as optional. Making the
last (`$out`) argument required allows the error middleware to execute.

Additionally, I noticed during step-debugging that `pipeErrorMiddleware()` was
calling `$this->pipe()` — which contains much of the same logic for preparing
lazy middleware. As such, I updated it to call `parent::pipe()` as a minor
optimization.