Added documentation deployment toolchain
This is the documentation deployment toolchain for generating gh-pages
on-the-fly from the master branch.

The Travis-CI configuration now defines several environment variables for
specifying the repository, github username and email, and github token
(encrypted) to use when committing the files.

`doc/deploy.sh` performs the actual build process. It first initializes a bare
repository under `doc/html/`, adds the repository as a remote, fetches it, and
resets head to the gh-pages branch. It then builds the documentation, adds all
changes, and pushes them back to the gh-pages branch.