Implemented landing page
This patch implements a landing page for the documentation. To do so, it
replaces the contents generated by mkdocs with bespoke contents that detail
marketing features and how to get started.

This patch also iterates through all generated HTML pages to add the
`class="img-responsive"` attribute to any images detected, ensuring that the
layout is consistent across all pages.