Moved build/deploy scripts to zf-mkdoc-theme
This commit removes the build/deploy scripts from the repository, and moves them
into the zf-mkdoc-theme repository, which includes them as of 0.2.0.

`.travis.yml` was updated to conditionally download and extract the latest
zf-mkdoc-theme archive, and to execute the build chain from it; it is also now
doing so from the repo under the zendframework organization.