Correct the documentation for configuring a container-driven container
I'm not sure if the documentation was correct previously, but whenever
I've tried to follow it, to the letter, it's resulted in segmentation
faults or the routes never being loaded. Given that, after some further
research, I believe that the previous recommendation was never going to
work. So, I've migrated some of the configuration from a setup using the
skeleton installer and updated the docs to reflect that.