Update delegator examples for zend-servicemanager v3
Updated the examples in the "route specific pipeline" recipe to
demonstrate both zend-servicemanager v2 and v3 support. Fixed the
configuration example to use the key `delegators` instead of
`delegator_factories`.

Also updated the zend-form delegator factory examples to ensure that the
docblocks used mirror those in the route specific pipeline, for
consistency.