Do not rely on Stratigility `getOriginal*()` methods
These are deprecated starting with Stratigility 1.3. This patch updates
the `TemplatedErrorHandler` and the `WhoopsErrorHandler` to instead use
the request `getOriginal*` attributes, which are injected in the
Stratigility request/response decorators, and which can be injected by
the `OriginalMessages` middleware as well.

Additionally, this updates the `Application::run()` method to ensure
that the `originalResponse` attribute is injected, as it is not in
earlier Stratigility versions.