Prevent deprecation notice based on triggering error middleware
Users cannot yet set the raiseThrowables flag, as there are not shipped
facilities for handling errors using templates or whoops. As such,
currently, if the queue is exhausted or an exception is thrown, error
middleware is triggered. This patch adds an error handler to
`Application::__invoke()` to swallow deprecation notices due to
invocation of error middleware.