Allow HEAD and OPTIONS requests for any matched route
Updates zend-expressive-router dependency to 1.3.2+, and related new
versions of router implementations to those supporting that version.

This patch adds new optional middleware, `Zend\Expressive\Middleware\ImplicitHeadMiddleware`
and `ImplicitOptionsMiddleware`,  that can be placed between the routing
and dispatch middleware in order to handle HEAD and OPTIONS requests
when they are only implicitly supported by the matched route.  Each only
triggers on the request method they are interested in, and then only
when:

- a route result is present in the request
- the route result contains a route
- and the route indicates the method is only implicitly supported

For the implicit HEAD support, it will also check to see if GET is
supported before returning a response; if so, it dispatches the next
middleware layer, providing an updated request using GET as the method,
and ensures that the response returned has an empty body.

This patch also incorporates tests from #398, with updates for the
following:

- Response bodies for implicit HEAD and OPTIONS requests should be empty.
- zend-expressive-aurarouter failed until version 1.1.3.