Updated to zend-expressive-router 2.0
Updated the depedency on zend-expressive-router to `^2.0`, and, in
development, each router implementation to that same constraint.

Doing so has an intersting side-effect: the route result is guaranteed
to have valid middleware composed, as the result is created from a
`Route` instance, which will raise an exception much earlier in cases of
invalid middleware being passed during its own instantiation. This
allowed removing two unit tests, as well as a conditional within
`dispatchMiddleware()` for verifying we have callable middleware.