Ensure Application and Integration tests work
Primarily went through and used interop middleware instead of callable
middleware, as assertions were checking that middleware composed in
routes were the same as the middleware passed to `pipe()` - callable
middleware gets decorated, which made these assertions fail, even though
the intent was correct.

Additionally, noted that the default delegate should return a default
404 response instead of returning an empty 200 response, and updated the
logic in `run()` to do so. In doing so, also realized that the
`NoopFinalHandler` does not accept a response instance, so the logic of
`getDefaultDelegate()` could be simplified.