Better server request exception handling
@webimpress noted that the `$exception` argument to
`emitMarshalServerRequestException()` was not being used, and I realized
that this is something that should be updated to provide better
reporting to end users. I accomplished this by checking for an
`ErrorResponseGenerator` in the container, and, if present, invoking it
with the exception, an empty request, and the response prototype.

If no container is present, or it does not have the
`ErrorResponseGenerator`, then the method simply uses the response
prototype to create a response with a 400 status to emit.