Change default delegate to be a DelegateInterface instance
Instead of using a callable with a convention-defined signature, expect
a `DelegateInterface` instead. For backwards compatibility, we will
still accept callable delegates, which will allow usage of things like
the `FinalHandler` or `NoopFinalHandler`, but these will be decorated
using `CallableDelegateDecorator` internally; we will eventually stop
doing that.

The `ApplicationFactory` now no longer looks for the `FinalHandler`
service, but instead the `DefaultDelegate` service.