Require http-interop/http-middleware 0.4.1
Since the newest versions of zend-stratigility and
zend-expressive-router each now depend on
webimpress/http-middleware-compatibility, http-interop/http-middleware
is no longer a required package. As such, new installations fail, as
this package is necessary in order to run Expressive.

This patch adds the package with a constraint of ^0.4.1 in order to fix
issues on the existing 2.0.X branch.