Provides more robust is_callable logic
Provides a new method, `IsCallableInteropMiddlewareTrait::isCallable()`,
which runs its argument through `is_callable()`. If the value passes, it
then checks to see if it is an array, and, if so, if:

- the first argument is an object or valid class name, AND
- the second argument is a valid method of the first argument.

This is done because `is_callable()` returns a false positive when a
two-element array is passed where the first argument is an object and
the second a string; the function does not verify that the second
element is a valid method of the first.