Extracts middleware creation to a new class, MiddlewareFactory
The methods `marshalMiddleware()`, `createPathDecorator()`, and
`createPipelineDecorator()` were extracted to a new class,
`MiddlewareFactory`. This class accepts a PSR-11 `ContainerInterface`
which it decorates as a `MiddlewareContainer`. It exposes a number of
methods for creating `MiddlewareInterface` instances, including:

- `prepare()`, which examines the type provided in order to delegate
  decoration to another method.
- `lazy()`, which decorates `string` middleware as
  `LazyLoadingMiddleware` instances.
- `callable()`, which decorates `callable` middleware as
  `CallableMiddlewareDecorator` instances.
- `pipeline()`, which passes the arguments it receives to `prepare()`
  and then pipes each to a `MiddlewarePipe` instance it returns.
- `path()`, which uses `prepare()` on its middleware argument in order
  to create and return a `PathMiddlewareDecorator`.

`Application` now creates a `MiddlewareFactory` during instantiation,
and uses it within `pipe()` and the various routing methods to ensure
a valid middleware is created for the pipeline and/or router.

This approach allows these methods to accept all the same middleware
types as previous major versions, with the exception of double pass
middleware and path-segregated middleware.