Creates ApplicationRunner
The `ApplicationRunner` composes a `RequestHandlerInterface`,
`EmitterInterface`, a callable server request factory, and a callable
error response generator (to invoke if the server request factory raises
a throwable or exception).

If no request is passed to `run()`, the class uses the composed server
request factory to generate a request. If an error occurs, it uses the
error response generator to create a response, which it then emits.

The request is passed to the composed handler, and the response returned
by it is then passed to the emitter.