Refactors integration tests
I determined the `ZendTest\Expressive\IntegrationTest` to be obsolete;
it's testing workflows that are tested elsewhere.

The `ZendTest\Expressive\Router\IntegrationTest`, however, was an
interesting testbed in terms of seeing how the middleware pipeline,
route middleware/router, and dispatch middleware work together,
particularly in terms of the supported router _implementations_. I
modified the test setup to create a `Proxy` instance instead, to see if
tests continued to pass. They did!