Allow using PSR-15 request handlers
This patch updates the `MiddlewareContainer` and `MiddlewareFactory` to
allow them to detect PSR-15 request handlers, and decorate them using
zend-stratigility's `RequestHandlerMiddleware`. This ensures type-safety
when used in either pipelines or with routed middleware, but also
ensures we can use the full-spectrum of available handlers with the
system.

A new method was also added to `MiddlewareFactory`: `handler()`. This
method accepts a `RequestHandlerInterface` instance and returns a
`RequestHandlerMiddleware` instance decorating it, making it simple to
use request handlers with the explicit functionality of
zend-expressive-router's `PathBasedRoutingMiddleware` API.