Renames NotFoundMiddleware to NotFoundHandler
Renames `Zend\Expressive\Middleware\NotFoundMiddleware` to
`Zend\Expressive\Handler\NotFoundHandler`. This is done to allow usage
in more contexts. Since it was already acting like a handler, by having
it explicitly be one, we can use it as one. The MiddlewareFactory will
decorate it as middleware when piped, allowing it to be used in that
context as well.