Adds constants for virtual services
Per the discussion on zendframework/zend-expressive-skeleton#220, this
commit converts virtual service names to constants. In each case, the
original resolved value of the virtual class name is used as the
constant value, but the config provider and factories now refer to the
constants instead.

Additionally, the `RouterResponseInterface` constant was discovered to
be no longer needed, due to the extraction of route-related middleware
to the zend-expressive-router package.