Use the `Zend\Expressive\ROUTE_MIDDLEWARE` to retrieve the path-based routing middleware.
In order to make upgrading from v2 easier, we need to ensure that we use
the same service name in the application factory as we use in the
bootstrap. Since changing the bootstrap requires manual intervention,
using the `Zend\Expressive\ROUTE_MIDDLEWARE` constant will present a
simpler upgrade, as it resolves to a string matching the v2 middleware
service name.