Versions documentation
This patch versions the documentation.

V1 docs were imported from the v1.1.1 tag, and pushed under a v1
subdirectory of `docs/book/`. A sibling `v2/` subdirectory was also
created, and the existing v2 docs were moved under that directory. The
v1-specific migration documentation was removed from the v2 tree,
leaving us with a single migration document for that tree.

v2 documentation was scrubbed of references to "version 1" and/or "since
v2"; users can go to the v1 documentation for that information.

All existing pages were updated to render as redirects, using a
combination of `<noscript>` tags and JS redirects.

The mkdocs configuration was updated:

- to rewrite all existing mappings to the v2 subdirectory
- to add links for the v1 docs
- to add hidden mappings for the existing redirect pages

I've tested builds locally, and all works as expected. Existing URIs
redirect to the correct v2 variants, and the navigation now uses v2
links.