Backports ApplicationConfigInjectionDelegator from release-3.0.0
This patch backports the ApplicationConfigInjectionDelegator and
modifies it to work with v2. The ApplicationConfigInjectionTrait was
updated to delegate to the new delegator. The ApplicationFactory now
also uses the delegator methods instead of the application methods.

Test changes were primarily due to usage of callable double-pass
middleware; updating them to use interop middleware simplifies setup and
prevents unnecessary deprecation notices.