Have `Application` compose a `RouteCollector` instead of `PathBasedRoutingMiddleware`
This is a change proposed while updating the documentation for version 3.
Having two routing middleware is confusing, particularly since the
`PathBasedRoutingMiddleware` simply provides methods for modifying the
state of the the composed router.

zend-expressive-router 3.0.0rc5 updates to rename
`PathBasedRoutingMiddleware` to `Zend\Expressive\Router\RouteCollector`,
and to no longer extend `RouteMiddleware`. This allows it to do its one
job: create and return `Route` instances that are also injected into the
composed router.

This patch updates to that version, and modifies `Application` to
compose a `RouteCollector` instead of `PathBasedRoutingMiddleware`.