Do not cast request handlers implementing middleware to `RequestHandlerMiddleware`
Per #645, `MiddlewareContainer::get()` was incorrectly casting
middleware that also implemented `RequestHandlerInterface` to
`RequestHandlerMiddleware`. This particularly affected
`Zend\Stratigility\MiddlewarePipe` instances, as they implement both
interfaces; if the last middleware in a pipeline called on the handler,
users would then encounter a "pipeline exhausted" error, because it
would be invoked as a handler, and thus have no reference to the
application's pipeline.

This change verifies that the request handler does not also implement
`MiddlewareInterface` before casting.